+++ GenerateBoundaryPairs/MAIN_GenerateBoundaryPairs.cpp
@@ -24,7 +24,7 @@ void TestLatticeTypes() {
    generator.randSellingReduced();  // try to get a random lattice that gives valid unit cells for all 
    const S6 temp = generator.randSellingReduced();
 
-   for (unsigned long i = 0; i < vDeloneTypes.size(); ++i) {
+   for (size_t i = 0; i < vDeloneTypes.size(); ++i) {
       const S6 test = vDeloneTypes[i] * temp;
       const bool b = LRL_Cell_Degrees(test).GetValid();
       //std::cout << LRL_Cell_Degrees(test) << std::endl;
@@ -33,13 +33,13 @@ void TestLatticeTypes() {
 
 int main()
 {
-   const unsigned long targetReducedUnreducedPairCount = 2000;
+   const size_t targetReducedUnreducedPairCount = 2000;
    GeneratePairsOfTestData::MultiGeneratePair_Reduced_Other_ReducesFarAway(targetReducedUnreducedPairCount);
  
-   const unsigned long targetRandomCount = 10000;
+   const size_t targetRandomCount = 10000;
    GeneratePairsOfTestData::GenerateRandomLatticesAndCatalogByReductionStepCount(targetRandomCount);
 
-   const unsigned long targetNearDeloneType = 1200;
+   const size_t targetNearDeloneType = 1200;
    GeneratePairsOfTestData::GenerateLatticeTypeExamplesNearDeloneTypes(targetNearDeloneType);
 
    TestLatticeTypes();
