+++ LatticeRepresentationLib/D7.h
@@ -56,11 +56,11 @@ public:
    bool operator== (const D7& d7) const;
    bool operator!= (const D7& d7) const;
 
-   double operator[](const unsigned long n) const { return m_vec[n]; }
-   double& operator[](const unsigned long n) { return m_vec[n]; }
+   double operator[](const size_t n) const { return m_vec[n]; }
+   double& operator[](const size_t n) { return m_vec[n]; }
    double DistanceBetween(const D7& v1, const D7& v2);
    std::vector<std::pair<std::string, std::string> > ClassifyVector(const double delta) const;
-   unsigned long size(void) const { return 7; }
+   size_t size(void) const { return 7; }
    double norm() const { return m_vec.norm(); }
    double norm(const D7& d7) const { return d7.norm(); }
    double Norm() const { return m_vec.norm(); }
@@ -73,7 +73,7 @@ public:
    bool IsValid(void) const { return m_valid; }
    void SetValid(const bool b) {m_valid = b; }
 
-   double at(const unsigned long n) const { return m_vec[n]; }
+   double at(const size_t n) const { return m_vec[n]; }
 
    std::pair<std::string, std::string> SummarizeBoundaries(const double delta) const;
 
@@ -89,7 +89,7 @@ public:
 
 private:
    VecN m_vec;
-   unsigned long m_dim;
+   size_t m_dim;
    bool m_valid;
 };
