+++ LatticeRepresentationLib/LRL_Cell_Degrees.cpp
@@ -14,7 +14,7 @@ LRL_Cell_Degrees::LRL_Cell_Degrees(const LRL_Cell& cell) {
    m_cell = cell.GetVector();
    m_valid = cell.GetValid() && m_cell[3] < pi && m_cell[4] < pi && m_cell[5] < pi && (m_cell[3] + m_cell[4] + m_cell[5])< twopi;
    if (m_valid)
-      for (unsigned long i = 3; i < 6; ++i)
+      for (size_t i = 3; i < 6; ++i)
          m_cell[i] *= 180.0 / (4.0*atan(1.0));
    else
       *this = LRL_Cell(0.0, 0.0, 0.0, 0.0, 0.0, 0.0);
@@ -24,7 +24,7 @@ LRL_Cell_Degrees::LRL_Cell_Degrees(const LRL_Cell& cell) {
 std::ostream& operator<< (std::ostream& o, const LRL_Cell_Degrees& c) {
    std::streamsize oldPrecision = o.precision();
    o << std::fixed << std::setprecision(5);
-   for (unsigned long i = 0; i < 6; ++i)
+   for (size_t i = 0; i < 6; ++i)
       o << std::setw(9) << c.GetVector()[i] << " ";
    o << std::setprecision(oldPrecision);
    o.unsetf(std::ios::floatfield);
